requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
psycopg2-binary==2.9.1
paho-mqtt==1.6.1
requests==2.26.0
utils~=1.0.1